pyproject.toml
[tool.poetry]
name = "code_analyzer"
version = "0.1.0"
description = "Code Analyzer"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
matplotlib = "^3.10.0"
pyyaml = "^6.0.2"
pytest = "^8.3.4"
cryptography = "^44.0.0"
seaborn = "^0.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
